Correctly handle component bounds for shadings 4-7
Follow-up to https://pdfium-review.googlesource.com/c/pdfium/+/132471

As that CL description says, CPDF_MeshStream::ReadColor() returns a
value in Decode[4]..Decode[5]. We used to multiply that value by 255 to
get an index into the 1D texture. In practice, this range is 0..1, but
for e.g. fuzz inputs it isn't. So add the actual mapping that I had
forgotten to put in.

(Since the 1D texture exists in cpdf_rendershading.cpp, I felt that the
mapping code should be there too, instead of modifying ReadColor().)

Tweak the test to use 1..2 instead of 0..1 for the functions in the 2nd
column. Without the code change, that is enough to also assert on that
test file.

Bug: 421329279
Change-Id: Id92d90d645503d14e295a24c8083641011acafc3