Interpolate shadings in correct color space
We interpolated gouraud and patch colors in RGB,
instead of in the shading's color space as dictated
by the spec. (Especially visible for e.g. LAB color
spaces.) Now we interpolate in the right color space
and only convert to RGB when painting a primitive.

CPDF_MeshStream::ReadColor() no longer returns an
RGB value but all components. The rest is mechanical
fallout from this.

XXX words on indexed handling

XXX CMYK perf

Bug: 406662638
Change-Id: I05616987674da10db90deb73ae75a95688d264f8