Treat all Unicode cmaps as MSUnicode
PDF 32000:2008 9.6.6.4 Encodings for TrueType Fonts outlines how to
handle TrueType with /Encoding. In this description only the handling of
(3,1) Unicode cmap is mandated. When nothing applies "a conforming
reader may supply a mapping of its choosing." Pdfium currently chooses
the encoding in the /ToUnicode.

Change this behavior to consider all Unicode cmaps, which includes
(3,1), (3,10), and (0,*). The /ToUnicode encoding is still present as a
fallback if no Unicode (or MacRoman) cmaps are present.

Bug: 378932943
Change-Id: Id22e3133d926efe52ec843257e3c1128f6a26ff1