Test that all TrueType Unicode cmaps are equal
A few parts of the PDF specification specifically call out TrueType cmap
(3,1) as the Unicode cmap. However, from a user perspective the format
of the cmap table is irrelevant, the only thing that matters is that the
cmap converts Unicode code points into glyph ids.

Add a test that ensures cmap (0,3) is treated the same as cmap (3,1) in
the case where a /Encoding /MacRomanEncoding and (an inaccurate)
/ToUnicode is present.

Bug: 378932943
Change-Id: Ifc82c51e3fa06c0ad1c559468b26c3bf8fc88689