Ensure all TrueType Unicode cmap formats are equal
A few parts of the PDF specification specifically call out TrueType cmap
format (3,1) as the sole TrueType Unicode cmap format. However, from a
user perspective the format of the TrueType cmap table is irrelevant,
the only thing that matters is that the character map converts Unicode
code points into glyph ids.

Add a test that ensures TrueType cmap format (0,3) is treated the same
as TrueType cmap format (3,1) in the case where a /Encoding
/MacRomanEncoding and (an inaccurate) /ToUnicode is present.

Bug: 378932943
Change-Id: Ifc82c51e3fa06c0ad1c559468b26c3bf8fc88689