Avoid loose bounds for invalid char codes
The current loose bounds computation assumes that the
CharInfo::char_code_ is valid. When the char code isn't valid the loose
bounds width will be zero.

In addition, the /ActualText kPiece currently attempt to store a valid
char code. However, this is incorrect as the text did not originate from
any char code. An incorrect char code could lead to the loose bounds
being calculated from a completely different glyph than that which was
used for drawing.

Bug: 42271607
Change-Id: Ie2d2d41287322f3aa3b28192d72f3a8301577d91