Fix a JPEG2000 decoding regression
When neither the PDF nor the JPEG2000 image provides a colorspace, treat
images with 3 channels as though they are RGB and decode them, instead
of giving up and not decoding at all.

This fixes a regression from CL [1], and also reverts the change to
testing/resources/pixel/bug_304_expected.pdf.0.png from that CL.
Although Acrobat Reader rejects the image in the top-left quadrant and
renders it as white, most other PDF implementations will render is as
red.

[1] https://pdfium-review.googlesource.com/104031

Bug: 430651971
Change-Id: I1bc40492671bd0e2bd39cab063d30157dce92e63