Split TextObject::Item rect for each code point
When a CharCode maps to multiple code points, the item is represented as
text by a sequence of CharInfo. Prior to this change each CharInfo was
assigned the full rectangle of the original CharCode.  This change
splits the the original CharCode's rectangle between the CharInfos. This
allows for each CharInfo to be independently selectable instead of all
of them overlapping.

Bug: 42271082
Change-Id: Ib9a9cde9ace0ab64c0e3a6ac86ae1aaf4b083881