Page Movement Fixes in XFA based PDFs in PDFIUM.
Fix page movement issues in PDFIUM, especially in XFA based PDFs as
mentioned in issue 433689235, made some fixes to the suggested patches
and ran tests with existing cases to ensure previous functionality stays
similar.

Credit for the original patch, and most code to Stefan Ziegler. The code
was slightly changed and tested with prior test cases
(perticularly pertaining to page movement) to ensure proper functionality.

TODO - Add an XFA specific testcase to test if pages move successfully
for XFA based pdfs, which can be useful for future bugs in this space.
Further, XFA testing also needs to be conducted to verify the fix.

Bug: 433689235
Change-Id: Ifc6ad58d149cfb9ec31d3cdb12dc761ef57493ec