Page Movement Fixes in XFA based PDFs in PDFium.
Fix page movement issues (with FPDF_MovePages() in XFA enabled PDFs) in PDFium. As mentioned in issue 433689235. This was done by ensuring different handling for the XFA extension and XFA-enabled PDFs to ensure that page movement functionality remains the same on these PDFs.

Fixes were made to the patch and tests were ran to ensure previous functionality stays similar.

Credit for the original patch, and most code to Stefan Ziegler. The code
was slightly changed and tested with prior test cases
(particularly pertaining to page movement) to ensure proper functionality.

TODO - Add an XFA specific test case to test if pages move successfully
for XFA based PDFs, which can be useful for future bugs in this space.
Further, XFA testing also needs to be conducted to verify the fix.

Bug: 433689235
Change-Id: Ifc6ad58d149cfb9ec31d3cdb12dc761ef57493ec