Change UseTTCharmapUnicode to prefer MSSymbol
Prior to "Treat all Unicode character maps like MSUnicode" [0] only the
cmap(3,1) (Windows,Unicode BMP) was considered as Unicode. After that
change, all Unicode cmaps like cmap(0,*) (Unicode,*) are considered as
Unicode. This fixes a number of issues with embedded fonts that do not
have a cmap(3,1).

However, this caused a regression in the case of embedded fonts with
Unicode cmap that isn't cmap(3,1) which also have a cmap(3,0)
(Windows,Symbol). In this case the Unicode cmap is now chosen over the
Symbol cmap. The Symbol cmap has special rules to augment the cmap as
well as special rules around converting to Unicode which will not be
applied to a Unicode cmap.

This changes the rule to prefer cmap(3,1) as before, then any other
Unicode cmap so long as there is no cmap(3,0). This restores the
previous behavior while still using Unicode cmaps when they are
available.

[0] https://pdfium.googlesource.com/pdfium/+/5da0e90faa40079145dc146f70bf186d3ad5e93f

Bug: 438810722
Change-Id: I247bb739fcb39b177923a7d4f3821df70dbca4fe