Use constexpr-if to avoid ugly code where possible
Only gotcha is all arms of the constexpr-if must be valid on
all platforms. The benefit is that we can catch cross-platform
compilation errors on any platform.

Part of the motivation is to better enable automatic unsafe
buffer rewrite tools (after applying this to other repos), since
unsafe buffers warnings are produced for both arms of the if,
even though one will be discarded.

Change-Id: Iebf53054ff67ec7a07072a22b7a180a189f05bf1