[rust png] Introduce `libpng`-agnostic `enum class EncodingColorType`.
This CL refactors the type of the `color_type` parameter of the
`fxcodec::PngDecoder::Delegate::PngReadHeader` method so that the API is
more `libpng`-agnostic (meaning that implementing the interface in
`ProgressiveDecoder` doesn't require reaching for `PNG_COLOR_TYPE_...`
constants from `png.h`, nor hardcoding values of those constants such
as `2` or `6`).

To simplify further refactorings, this CL also identifies some
unreachable `switch` `case`s and then collapses them into a single
`NOTREACHED` `case` with a comment.  This is possible, because there is
currently only one implementation of the `PngDecoder::Delegate`
interface, and this implementation only ever asks to decode into
`Truecolor` or `TruecolorWithAlpha` format.  The unreachable cases are
further refactored away in a follow-up CL here:
https://pdfium-review.googlesource.com/c/pdfium/+/135615

Bug: 444045690
Change-Id: I588b47d0d01b11eb528b56d3394eb2c962cfc618