Fix selecting replaced text after undo
Pdfium uses a stack of undo items to implement undo/redo. For text
replacement there are a variable number of undo items added to stack.
The UndoReplaceSelection undo item class was created in
https://pdfium.googlesource.com/pdfium/+/3abc96a18bf41a55958e2fe99a3ce03882c3ecb5

I believe the SelectNone calls that caused this bug were simply a
copy-paste error since many of the other undo items call that.
UndoReplaceSelection is only a sentinel item that tracks the number of
intermediate undo items, it shouldn't do any actions itself.

Bug: chromium:445215584
Change-Id: Ife7f13aca9dbe2260f2f4a1b2ef5650099f44dd5