Improve formfill_embeddertest failure messages
Because of the way the helpers were written, all test failures just
showed the line of the EXPECT that failed in the helper function, not
the line in the test that failed. So if you get a test failure you have
to edit the code to figure out what happened (potentially using
SCOPED_TRACE).

Instead, we should use gtest APIs to make our helper functions. I
considered using a custom matcher but matchers are required to be pure
functions and most of these helpers have side effects. We could also
solve this problem with a series of macros that use __LINE__ and
__FILE__ or SCOPED_TRACE instead of calling the functions but that
seemed awkward given that these are test fixture helpers that need the
member variables. We also could have used EXPECT_PRED* macros and made
our functions return bool instead of using EXPECT macros internally but
that didn't seem like a nice user-interface.

Here's tho docs on this
https://google.github.io/googletest/advanced.html#predicate-assertions-for-better-error-messages

Change-Id: I3d7985c7a460f873c9242a003ee79ff7ca4ece7e