Fix selecting replaced text after undo
PDFium uses a stack of undo items to implement undo/redo. For text
replacement there are a variable number of undo items added to stack.
The UndoReplaceSelection undo item class was created in
https://pdfium.googlesource.com/pdfium/+/3abc96a18bf41a55958e2fe99a3ce03882c3ecb5

I believe the SelectNone calls that caused this bug were simply a
copy-paste error since many of the other undo items call that.
UndoReplaceSelection is only a sentinel item that tracks the number of
intermediate undo items, it shouldn't do any actions itself.

For ReplaceAndKeepSelection arguably when the redo action runs the
replaced text should be selected as it was originally, but in this
change it is left unselected like after redoing ReplaceSelection. I
think this makes sense because in general selection is not an undo item,
pasting over a selection is a special case that is widely implemented.
If you select text, then press undo then redo, the text will not be
selected anymore. There is no equivalent operation of "paste and then
select the pasted text" in other programs, so nothing to defer to. Also
this is an experimental API for pdfium. So overall I think it makes
sense to leave as-is instead of adding a new undo item type for this
selection.

Fixed: 445215584
Change-Id: Ife7f13aca9dbe2260f2f4a1b2ef5650099f44dd5