Improve CFX_SubstFont font name value
CFX_FontMapper::UseExternalSubst() tries to talk to the external font
interface to get a font name. If this fails, sometimes the CFX_SubstFont
name may end up being the same as the /BaseFont value. This confuses
IsActualFontLoaded() in core/fpdfapi/render/charposlist.cpp, which is
trying to apply some heuristics to see if the external font interface
got a font that matches the /BaseFont.

To fix this confusion and make the CFX_SubstFont font name better,
update UseExternalSubst() to read the actual font name from the
substitute font data.

- This does not require any changes in the embedder that implements the
  external font interface.
- This is likely what the external font interface would do anyway, but
  doing this in UseExternalSubst() creates very little extra work, as it
  already has to load the substitute font.

Bug: 445171006
Change-Id: Ia3f0f7f3e8a1af03fa3551b658ec9d3adce13dca