Don't handle edit shortcuts in FORM_OnKeyDown()
In order to get fully correct behavior, it is necessary to use the
Chrome platform-specific handler for these shortcuts on MacOS. On other
platforms currently most of these shortcuts are handled in
FORM_OnChar(). Also the chrome OnKeyDown handler for these shortcuts
only happens on MacOS. So this change will have no effect on other
platforms.

Lower in the call stack, FORM_OnKeyDown() is hard-coded to return true
when the key event is passed off to the widget message queue in
CXFA_FFField::OnKeyDown(). This means that CPDFSDK_Annot::OnKeyDown()
always returns true (when a form field is focused, links can return
false). So we need to filter out the events we want to return false for
before that function call.

When I simplified the logic in CPDFSDK_PageView::OnKeyDown() I slightly
changed the behavior. Before this CL it was possible to block tab focus
change handling if you pressed Ctrl+tab or alt+tab and that event would
be passed to CPDFSDK_Annot::OnKeyDown(). However, none of the downstream
classes actually handled this event other than returning true and doing
nothing.

After this CL all of the shortcut keys work on MacOS but the edit menu
bar does not indicate the action for all of the shortcuts. There's some
handler cleanup work to do still in chrome to fix that.

Also in this CL I fixed a small issue where Cmd+Tab was handled the same
as just Tab which Ctrl+Tab and Alt+Tab are ignored on other platforms.

Bug: 413695643
Change-Id: I031d58f8d131bb342daf8c2594f2b69f427289ec