Remove dead code in xfa/fwl/cfwl_notedriver.cpp
Technically it was possible to trigger this code path if you turn on the
XFA setting in chrome://flags, open an XFA text field PDF
(testing/resources/xfa/xfa_multiline_textfield.pdf), focus the field,
and press ctrl+tab or alt+tab on a non-macos platform. However this code
path doesn't do anything with that event. These are system shortcuts
that are often handled before the event makes it to chrome anyway.

Originally there was some code in this branch but it was removed in 2020
in https://pdfium.googlesource.com/pdfium/+/f0c6ea45cb0c0500956093a693f2fddc2013ff9f

I'm not 100% sure what the intention was for this code. It seems like it
was meant to handle focus changing on tab press but that's now handled
in CPDFSDK_PageView::OnKeyDown()

Change-Id: I39ecd32bb735bd6d841f3764a3809f28e463f677