Refactor CPDFSDK_PageView::OnKeyDown()
I had looked at this while fixing a bug and thought the logic could be
simpler so that there are not two separate code blocks checking for tab
key handling.

There's one small behavior change: Before this CL it was possible to
block tab focus change handling (the section of code that calls
GetNextAnnot()) if you pressed Ctrl+tab or alt+tab and that event would
be passed to CPDFSDK_Annot::OnKeyDown(). However, none of the downstream
classes actually handled this event other than returning true and doing
nothing.

Bug: 413695643
Change-Id: I031d58f8d131bb342daf8c2594f2b69f427289ec