Don't block edit shortcuts in CPWL_Edit::OnKeyDown
The IsProceedToOnChar function has been around since the initial PDFium
import:
https://pdfium.googlesource.com/pdfium/+/5110c4743751145c4ae1934cd1d83bc6c55bb43f/fpdfsdk/src/pdfwindow/PWL_Edit.cpp#956

Based on the comments and function name it seems to me that the purpose
of this code was to intentionally eat these OnKeyDown events because
these shortcuts are handled in OnChar. Now, particularly in Chrome, this
logic blocks MacOS shortcut handlers when PDFium eats these events.

This only fixes the cmd+z issue in Chrome for non-XFA forms since this
return true was only in the non-XFA textbox class.

Bug: 413695643
Change-Id: Icb7bcc6c5f1f914c5999eba68a26211f74f45f3f