Don't block edit shortcuts in CXFA_FFField::OnKeyDown
Since the events for XFA fields are passed into a message queue which
doesn't have a way to return the whether or not the event was handled,
there is a hard-coded return true here. So in order to not handle the
shortcut keys so that embedders can handle them, it is necessary to put
in a filter here. None of the types handle these shortcuts, although
some handle things like ctrl+home.

This will make cmd+z work on MacOS in embedders for XFA fields, if the
embedders handle OnKeyDown and call FORM_Undo().

Bug: 413695643
Change-Id: I8b39c8f338ac02a3d9bda6230c6d25c6e250ff5b