Fix cmd+Home & cmd+End in textboxes
Before this CL these new tests would have passed except for on MacOS.
The two tests are the same except that in XFA forms there's no \r
inserted when return is pressed.

There is no PDFium API, as far as I know, to get the current caret
position in a form field. So I had to be creative and use delete and
backspace key presses to check the caret location.

Fixed both XFA and non-XFA forms.

Bug: 448699368
Change-Id: Ic5d535ace1b0bb88582351a775f41dd0acf6a410