[rust png] Replace `src_color_type` with `number_of_components`.
Before this CL, `ProgressiveDecoder::PngReadHeader` would only use
`src_color_type` to calculate `src_components_`:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/pdfium/core/fxcodec/progressive_decoder.cpp;l=99;drc=1a1c55244b379198ca0eaa9498c1fd527850abb2

After this CL, the signature of `PngReadHeader` is simplified, so that
the caller can directly pass the number of components, without the
indirection of having to convert to/from `EncodedColorType` enum.

This CL also expands some abbreviated field and variable names to help
with readability.

Bug: 444045690
Change-Id: Ida372cb2448cbf57b3f59a0dc6cc76fed54e2103