Remove dead code: CPWL_Edit::{Copy,Paste}Text()
The only behavior change is that now FORM_OnChar() returns false for
ctrl+c and ctrl+v, as well as FORM_OnKeyDown() returns false for insert
and shift+insert. This is correct because PDFium never has done anything
with those events.

Originally PDFium contained clipboard handling code which these
functions used:
https://pdfium.googlesource.com/pdfium/+/5110c4743751145c4ae1934cd1d83bc6c55bb43f/fpdfsdk/src/pdfwindow/PWL_Edit.cpp#170
However that clipboard handling code was deleted in 2016:
https://pdfium.googlesource.com/pdfium/+/b95901091d63ce09accec3088945955f969d46e1%5E%21/#F14

These exact functions deleted in this CL have been empty since the
initial import of PDFium:
https://pdfium.googlesource.com/pdfium/+/5110c4743751145c4ae1934cd1d83bc6c55bb43f/fpdfsdk/src/pdfwindow/PWL_EditCtrl.cpp#569
However initially they were virtual functions and that clipboard
handling code was the overload. The overloading was dropped in 2015 in
https://pdfium.googlesource.com/pdfium/+/2b1a2d528469cda4e9f3e36d3c7a649e0d476480%5E%21/#F35

So I think in 2016 when the clipboard handling code was deleted it was
only used by the context menu handling code (which used to be in PDFium
and now Chrome handles). At that time the keyboard shortcut handling
code in PWL_EditCtrl.cpp was empty functions, which I believe were
originally meant to be virtual function calls to the system clipboard
code that was in PWL_Edit.cpp.

Today PDFium in fact has no way to access the system clipboard.

Change-Id: Ifb9f73c8b51e5c3138c7c45d32df450c436d9e25