Delete dead code: CPWL_Edit::CutText()
The behavior difference is that now for some embedders shift+delete is
allowed in password fields by PDFium. In Chrome this already worked via
ExecuteEditCommand() anyway. Additionally now FORM_OnChar returns false
for ctrl+x. Lastly PDFium would internally delete the selected text on
ctrl+x previously, but this code path was never used by Chrome.

It's possible some embedders use that behavior but PDFium has not
actually put the text in the clipboard since at least 2016, possibly
even broken in 2015.

ClearSelection() internally checks if there is a selection or if the
field is read-only so those checks aren't needed.

Change-Id: Ie5f120b0292171caeb3c493ec804ec0483603f85