Fix double handling the delete key event
Previously for some embedders (Chrome) that send both OnChar and
OnKeyDown for delete PDFium was double handling the delete key. The
OnChar handling was unexpected and was inserting 0x7F as a character in
the text field, which caused weird behavior like inserting a * in
password fields but not actually having any text if you select all and
copy it.

This double handling inserted two items in the undo queue so when you
pressed delete you had to press undo twice.

In JavaScript events we only receive an OnKeyDown event for the delete
key so that seemed like the proper place to handle it.

There is a testing/resources/pixel/password.in file which we could have
generated a pdf for and set up a test fixture for. However I'm not
really sure how I would test that issue anyway. The test making sure we
don't handle OnChar delete should be enough to cover that issue.

Fixed: 454273254
Change-Id: Ibccc44b2f6792428eaf7e02439990cb5f55110c6