Fix double handling the delete key event
Previously for some embedders (Chrome Linux) that send both OnChar and
OnKeyDown for delete PDFium was double handling the delete key. The
OnChar handling was unexpected and was inserting 0x7F as a character in
the text field, which caused weird behavior like inserting a * in
password fields but not actually having any text if the user performs
select all and copies it.

This double handling inserted two items in the undo queue so when you
pressed delete you had to press undo twice.

In JavaScript code the browser only sends a "keydown" event for the
delete key so OnKeyDown seemed like the proper place to handle it in
PDFium to match the JavaScript event model.

There is a testing/resources/pixel/password.in file which I could have
generated a pdf for and set up a test fixture for. However I'm not
really sure how I would test the phantom * characters issue. The test
making sure PDFium doesn't handle OnChar delete should be enough to
cover that issue.

Fixed: 454273254
Change-Id: Ibccc44b2f6792428eaf7e02439990cb5f55110c6