Add crbug.com links to TODOs in testing/SUPPRESSIONS
Many of the bug IDs in the TODOs have the format pdfium:XXXX and

chromium:XXXXXX in testing/SUPPRESSIONS. CodeSearch does not support
this format, and newer PDFium developers may not know how to find the
bug report. Add crbug.com links to all bug IDs, and remove the
"chromium:" prefix where necessary.
Change-Id: I65a4a0d7628aca250f1f66719863492660e6dfc3