[rust png] Call `CFX_CodecMemory::Consume` on PNG input.
A WIP Skia-based codec implements `SkStream::read` on top of
`CFX_CodecMemory::ReadBlock` which "consumes" the bytes - the
`CFX_CodecMemory::pos_` field is updated.  This CL updates the legacy,
`libpng`-based decoded to also "consume" the bytes in a similar manner
(see the changes in `libpng_png_decoder.cpp`).

The change above exposed a problem in
`ProgressingDecoder::PngContinueDecode` which never reset `pos_` before
this CL.  This meant that the subsequent chunks of input were never made
available to the decoder as "unconsumed" bytes.  This is fixed by adding
a call to `Seek(0)`.

The final, opportunistic change in this CL renames a local variable from
`bResult` to `success` (following the current style guide).

Bug: 444045690
Change-Id: Ia44ce0bb7c923b6cd4adb4efe1f93f61755171d4