[Skia] Optimize DrawPath() when knockout is on
Instead of calculating geometric union and subtraction of paths which
may contain many points, composite the objects in pixel-space using
blend modes.

According to https://crbug.com/334774307#comment21 this implementation
should be non-lossy since the saveLayer system should allow the mask
resolution to scale and operations replayed.

Initially I kept the same structure of doing a union or difference based
on the stroke color and drawing the stroke on the separate layer from
the fill. However when I read ISO 32000-1:2008 I realized I could
simplify it and make the function more accurately follow the spec.

Rendering performance for the pdf in the bug using pdfium_test:

 - Before: 68.72s user 0.37s system 99% cpu 1:09.14 total
 - After: 0.52s user 0.12s system 99% cpu 0.645 total

There are some cases with single point paths where this code is
different from the previous implementation. However neither are correct
to the spec, and both draw all of the points the spec says to draw. I
will clean this up in a follow-up change.

Bug: 334774307
Change-Id: I7cc2d264bb06182c69d74dc7fcd04fe285d47256