Fork Skia Freetype code experiment
Add a new third_party/skia_freetype with a few files needed to support
Freetype usage when rendering with Skia. The files are copied over from
Skia. At the same time, change the Skia build configuration to avoid
building Skia with Freetype support.

With this shuffle, the end behavior is roughly the same, but the
FreeType dependency has been moved over from Skia to PDFium.

Change-Id: I37178878c5cc50bdb5d665b60e32615a4c569476