Add mac_xfa_x86 builder
This is a transitional step towards switching to ARM by default for
macOS. Currently, macOS builders default to x86, and mac_xfa_arm is the
one exception. In the near future, the default will be ARM and
mac_xfa_x86 will be the exception.

One difference is that mac_xfa_x86 is on CI, since it is known to work,
while mac_xfa_arm never was because it did not work until earlier today.

Bug: 41480180
Change-Id: I088deda81cc5fd6b70b2b3cedc8a05c1680098ea