Allow matching expected files in some cases in pngdiffer.py
https://pdfium-review.googlesource.com/c/pdfium/+/137611 updates test
expectations such that the en_14_foxit_products/2 test result on macOS
Skia matches AGG Linux, but not Skia Linux. However, updating the test
expectation to match AGG Linux causes the test to fail due to a matching
expectation. Removing the macOS Skia expectation would cause the test to
default to Skia Linux's expectation.

To fix this, soften the restrictions to allow matching expected files as
long as not all files match.

Change-Id: I5452d3c324c5d7c8c1bbc589fdbd66909583cfc9