Consolidate CFX_Face::Open() callers
This changeset consolidates the code in callers of CFX_Face::Open(). The
functions CFX_Font::LoadFile() and LoadFace() in cfgas_fontmgr.cpp are
two callers which share nearly identical logic. Though with fewer shared
aspects, android/cfpf_skiafontmgr.cpp also uses the consolidated code.

The code is consolidated into a function called OpenFromStream() and an
edited Open() function, with the former invoking the latter. These
functions are utilised in the callers to continue working toward the
goals of freetype encapsulation.

Bug: 452087741
Change-Id: Ic745d6b3c29b0736c38aaddafcc76e815976ac96