Fix progressive unoptimized decoding of JBIG2 generic region with template 1
ProgressiveDecodeArithTemplate0Unopt,
ProgressiveDecodeArithTemplate2Unopt,
ProgressiveDecodeArithTemplate3Unopt all use `loop_index_` for the
height loop. All the four Opt3() versions of these methods do that too.

ProgressiveDecodeArithTemplate1Unopt used to just use a local variable
`h` for that loop instead, like the non-progressive decoders further up
do. But ProgressiveDecodeArithTemplate1Unopt does have a return for
progressive encoding at the end of the height loop, which meant it keeps
processing scanline 0 over and over.

Fix to `loop_index_` as the height variable.

The test for this is in
https://pdfium-review.googlesource.com/c/pdfium_tests/+/137990

Bug: 459508595
Change-Id: I92770a82dbe13fd6ca6b8660631a55697453d579